LIS-65 Changing all database models to use decimal instead of float, because of accuracy issues.

+++ Sources/Common.WNP/Model/CurrentTransformerTestResult.cs
@@ -35,6 +35,6 @@ namespace AMSLLC.Listener.Common.WNP.Model
         /// <value>
         /// The test current (Amps).
         /// </value>
-        public float TestCurrent { get; set; }
+        public decimal TestCurrent { get; set; }
     }
 }
